feat(elements) Componente Admin con Alert
El componente Administradores ahora tiene el servicio de alerta y el toast

+++ reservas-asw-front/src/app/admin/pages/admins/table/list-admin.component.ts
@@ -2,6 +2,8 @@ import { Component, OnInit } from '@angular/core';
 import { AdminsService } from 'src/app/admin/services/admins.service';
 import { Administrador, AdminsResponse, AdminResponse } from '../../../interfaces/admin.interfaces';
 import { RouteName } from '../../../../../utils/enums';
+import { ToastsService } from '../../../../services/toasts.service';
+import { AlertsService } from '../../../../services/alerts.service';
 
 @Component({
   selector: 'app-list-admin',
@@ -15,18 +17,20 @@ export class ListAdminComponent implements OnInit {
   admins: Administrador[] = [];
 
   constructor(
-    private adminsService: AdminsService
+    private adminsService: AdminsService,
+    private toastService: ToastsService,
+    private alertService: AlertsService
   ) { }
-  
+
   ngOnInit(): void {
     this.getAdmins();
   }
-  
-  getAdmins(){
+
+  getAdmins() {
     this.adminsService.getAdmins()
-    .subscribe(
-      (adminsResponse: AdminsResponse) => this.admins = adminsResponse.data
-    );
+      .subscribe(
+        (adminsResponse: AdminsResponse) => this.admins = adminsResponse.data
+      );
   }
 
   get results(): Administrador[] {
@@ -36,10 +40,23 @@ export class ListAdminComponent implements OnInit {
   selectedAdmin!: Administrador;
 
   deleteAdmin(id: number): void {
-    this.adminsService.deleteAdmin(id)
-    .subscribe(
-      (adminResponse:AdminResponse) => this.admins = this.admins.filter((admin:Administrador)=> admin.idAdministrador !== id)
-    );
-    console.log(id);
-  };
+    this.alertService.showConfirmDialog({
+      message: '¿Desea eliminar el administrador? esta acción no se podrá revertir',
+      header: 'Eliminar Administrador',
+    })
+      .then(resp => {
+        if (resp) {
+          this.adminsService.deleteAdmin(id)
+            .subscribe(
+              (adminResponse: AdminResponse) => {
+                this.admins = this.admins.filter((admin: Administrador) => admin.idAdministrador !== id);
+                this.toastService.showToastSuccess({ summary: 'Administrador Eliminado', detail: 'El administrador ha sido eliminado correctamente' });
+              });
+          console.log(id);
+        } else {
+          return;
+        }
+      })
+      .catch(console.log);
+  }
 }
